refactor: Adiciona types em pasta segmentada
