+++ src/domain/interfaces/IBuilder.ts
@@ -1,8 +1,6 @@
-// domain/interfaces/IBuilder.ts
-
 /**
  * Contrato base para todos os builders do Beiju.
- * @template T — tipo produzido pelo build()
+ * @template T
  */
 export interface IBuilder<T> {
   build(): T
