+++ src/domain/model/clause/WhereCondition.ts
@@ -1,10 +1,10 @@
 import { ColumnRef } from "../ColumnRef.js";
-export type WhereOp = '=' | '!=' | '<>' | '>' | '<' | '>=' | '<=' | 'BETWEEN' | 'IN' 
+import { WhereOpType } from "../../types/WhereOpType.js";
 
 export class WhereCondition {
   constructor(
     readonly column: ColumnRef,
-    readonly op: WhereOp,
+    readonly op: WhereOpType,
     readonly value: unknown,
   ) {}
 }
