+++ src/domain/types/AggFnType.ts
@@ -1,0 +1,2 @@
+
+export type AggFnType = 'SUM' | 'AVG' | 'COUNT' | 'MIN' | 'MAX'
