+++ src/domain/types/QueryNodeKindType.ts
@@ -1,0 +1,9 @@
+export type QueryNodeKindType =
+  | 'SelectQuery'
+  | 'ColumnRef'
+  | 'AggregateExpr'
+  | 'WindowSpec'
+  | 'WindowFunctionExpr'
+  | 'WhereClause'
+  | 'OrderByItem'
+  | 'FrameSpec'
