+++ src/domain/types/SqlType.ts
@@ -1,0 +1,1 @@
+export type SqlType = 'number' | 'string' | 'date' | 'boolean'
