+++ src/domain/types/WhereOpType.ts
@@ -1,0 +1,1 @@
+export type WhereOpType = '=' | '!=' | '<>' | '>' | '<' | '>=' | '<=' | 'BETWEEN' | 'IN'
