+++ src/domain/types/WindowFnType.ts
@@ -1,0 +1,1 @@
+export type WindowFnType = 'ROW_NUMBER' | 'RANK' | 'DENSE_RANK' | 'LAG' | 'LEAD' | 'NTILE'
